Delete stop_binary, check that package is loaded before taking screenshot

+++ R/z_pages.R
@@ -48,27 +48,30 @@ animint2pages <- function(plot.list, github_repo, owner=NULL, commit_message = "
       stop(sprintf("Please run `install.packages('%s')` before using this function", pkg))
     }
   }
-  chrome.session <- chromote::ChromoteSession$new()
-  res <- animint2dir(plot.list, open.browser = FALSE, ...)
-  #Find available port and start server
-  portNum <- servr::random_port()
-  normDir <- normalizePath(res$out.dir, winslash = "/", mustWork = TRUE)
-  start_servr(serverDirectory = normDir, port = portNum, tmpPath = normDir)
-  Sys.sleep(3)
-  url <- sprintf("http://localhost:%d", portNum)
-  chrome.session$Page$navigate(url)
-  screenshot_path <- file.path(res$out.dir, "Capture.PNG")
-  screenshot_full <- file.path(res$out.dir, "Capture_full.PNG")
-  Sys.sleep(3)
-  ## Capture screenshot
-  chrome.session$screenshot(screenshot_full, selector = ".plot_content")
-  image_raw <- magick::image_read(screenshot_full)
-  image_trimmed <- magick::image_trim(image_raw)
-  magick::image_write(image_trimmed, screenshot_path)
-  unlink(screenshot_full)
-  chrome.session$close()
-  # Stop the server
-  stop_server(normDir)
+  
+  if(requireNamespace("chromote") && requireNamespace("magick")) {
+    chrome.session <- chromote::ChromoteSession$new()
+    res <- animint2dir(plot.list, open.browser = FALSE, ...)
+    #Find available port and start server
+    portNum <- servr::random_port()
+    normDir <- normalizePath(res$out.dir, winslash = "/", mustWork = TRUE)
+    start_servr(serverDirectory = normDir, port = portNum, tmpPath = normDir)
+    Sys.sleep(3)
+    url <- sprintf("http://localhost:%d", portNum)
+    chrome.session$Page$navigate(url)
+    screenshot_path <- file.path(res$out.dir, "Capture.PNG")
+    screenshot_full <- file.path(res$out.dir, "Capture_full.PNG")
+    Sys.sleep(3)
+    ## Capture screenshot
+    chrome.session$screenshot(screenshot_full, selector = ".plot_content")
+    image_raw <- magick::image_read(screenshot_full)
+    image_trimmed <- magick::image_trim(image_raw)
+    magick::image_write(image_trimmed, screenshot_path)
+    unlink(screenshot_full)
+    chrome.session$close()
+    # Stop the server
+    stop_server(normDir)
+  }
   all_files <- Sys.glob(file.path(res$out.dir, "*"))
   file_info <- file.info(all_files)
   to_post <- all_files[!(file_info$size == 0 | grepl("~$", all_files))]
